Change default_layer to dvorak
Overall I wanted consistency in chording keys (keystrokes composed of
modifiers), so Control, Alt, Shift, and Gui (Windows/Command/Mod4) are in mirrored
positions in the center of the thumbcluster.

This required keycodes I use less often to be moved from the thumbcluster harder to hit
switches or lower layers altogether.

- HOME and END are removed all together (though I add them to <mod>/<fn>)
- PG_UP and PG_DN are moved to left pinky, previously EQUAL + TAB
- TAB is moved to where DEL was as I use it for completion... alot
- DEL is moved to <keypad> layer, in the same place.
- EQUAL is moved to middle left finger down.. replacing CAPS